fix (updated seed to match new db models)

+++ server/db/models/product.js
@@ -8,7 +8,7 @@ const Product = db.define('product', {
     allowNull: false
   },
   price: {
-    type: Sequelize.DECIMAL
+    type: Sequelize.INTEGER
   },
   inventoryQuantity: {
     type: Sequelize.INTEGER,
